+++ tests/unit/small_text/test_imports.py
@@ -52,6 +52,7 @@ class ImportTest(unittest.TestCase):
     def test_import_query_strategies_module(self):
         from small_text import constraints
         from small_text import ClassificationType
+        from small_text import ClassBalancer
         from small_text import EmptyPoolException
         from small_text import QueryException
         from small_text import PoolExhaustedException